requirements.txt
streamlit
langchain
langchain-community
chromadb
sentence-transformers
pydantic
python-dotenv
google-generativeai